move sign/verify tests out to an example
